Move signup magic link fragment to login library
