+++ WordPress/src/main/res/values/strings.xml
@@ -1858,11 +1858,6 @@
     <string name="signup_epilogue_hint_password">Password (optional)</string>
     <string name="signup_epilogue_hint_password_detail">You can always log in with a magic link like the one you just used, but you can also set up a password if you prefer.</string>
     <string name="signup_epilogue_hint_username">Username</string>
-    <string name="signup_magic_link_error">There was some trouble sending the email.  You can retry now or close and try again later.</string>
-    <string name="signup_magic_link_error_button_negative">Close</string>
-    <string name="signup_magic_link_error_button_positive">Retry</string>
-    <string name="signup_magic_link_message">We sent you a magic signup link! Check your email on this device, and tap the link in the email to finish signing up.</string>
-    <string name="signup_magic_link_progress">Sending email</string>
     <string name="signup_updating_account">Updating account&#8230;</string>
     <string name="signup_user_exists">Email already exists on WordPress.com.\nProceeding with login.</string>
     <string name="content_description_add_avatar">Add avatar</string>
