+++ libs/login/WordPressLoginFlow/src/main/res/layout-land/signup_magic_link.xml
