+++ libs/login/WordPressLoginFlow/src/main/res/layout/signup_magic_link.xml
